google-trends-bigquery-pipeline/requirements.txt
google-cloud-bigquery
pytrends
pandas
numpy
matplotlib
python-dateutil